:recycle: refactors to add cards/decks/layout folders to components as component folder was getting crowded
